+++ bookscraper_backend/backend_api.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI, HTTPException
-from src.backend import (
+from bookscraper_backend.backend import (
     process_country_count,
     extract_authors,
     generate_country_count,
@@ -8,7 +8,7 @@ from goodreads_scraper.scrape import process_goodreads_url
 from pydantic import BaseModel, HttpUrl
 from fastapi.middleware.wsgi import WSGIMiddleware
 from werkzeug.middleware.profiler import ProfilerMiddleware
-from src.setup import setup_db
+from bookscraper_backend.setup import setup_db
 from graph_db import create_constraints
 from contextlib import asynccontextmanager
 from logger import logger
