+++ pyproject.toml
@@ -3,12 +3,13 @@ name = "bookscraper-backend"
 version = "0.1.0"
 description = ""
 authors = ["Tamir Einhorn Salem <tamtamireinhorn@gmail.com>"]
+packages = [{ include = "*", from = "src" }]
 readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.10"
 selenium = "^4.16.0"
-goodreads-scraper = {git = "https://github.com/BookExplorer/Goodreads-Scraper.git"}
+goodreads-scraper = { git = "https://github.com/BookExplorer/Goodreads-Scraper.git" }
 pandas = "^2.2.0"
 pycountry = "^23.12.11"
 sqlalchemy = "^2.0.27"
